refactor(tegg): fix asyncRun race, register streamRun in runningTasks, extract test helper
- Fix asyncRun race condition: register in runningTasks before IIFE starts
  executing to prevent the IIFE's finally block from deleting the entry
  before it is set
- Register streamRun in runningTasks so cancelRun/destroy can manage
  streaming runs; add cleanup in finally block
- Merge two sequential cancel store writes into one in streamRun abort path
- Extract MapStorageClient to shared test/helpers.ts to eliminate duplication
